Accumulates in __int128
* Any values >= LLONG_MAX or <= LLONG_MIN are considered nan
* execute_fixed returns LLONG_MAX if value is beyond, as unclear how to
push __int128 into python. py::int_() does the wrong thing
* Correctly detects energy overflows due to accumulation

+++ timemachine/cpp/src/gpu_utils.cuh
@@ -87,6 +87,6 @@ double __device__ __forceinline__ radd_rn(double a, double b) { return __dadd_rn
 
 void __global__ k_accumulate_energy(
     int N,
-    const unsigned long long *__restrict__ input_buffer, // [N]
-    unsigned long long *__restrict u_buffer              // [1]
+    const __int128 *__restrict__ input_buffer, // [N]
+    __int128 *__restrict u_buffer              // [1]
 );
